Cleaned test_xarray and added aoi, large_aoi to init
