+++ tests/__init__.py
@@ -1,5 +1,22 @@
 from __future__ import annotations
 
+import geopandas as gpd
+import pytest
+
 # import os
 # if not os.environ.get('MAXAR_API_KEY'):
 #    os.environ['MAXAR_API_KEY'] = 'fake-test-key'
+
+
+@pytest.fixture
+def aoi():
+    # 11 vertices, 1,361km^2
+    aoi_url = "https://raw.githubusercontent.com/SlideRuleEarth/sliderule-python/main/data/grandmesa.geojson"
+    return gpd.read_file(aoi_url)
+
+
+@pytest.fixture
+def large_aoi():
+    # 260 vertices, large area 269,590 km^2
+    aoi_url = "https://raw.githubusercontent.com/unitedstates/districts/refs/heads/gh-pages/states/CO/shape.geojson"
+    return gpd.read_file(aoi_url)
